% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{freqseq}
\alias{freqseq}
\title{Frequency nominal central bands}
\usage{
freqseq(from = 20, to = 20000, bands = "thirds")
}
\description{
Frequency nominal central bands
}
\examples{

freqseq()
freqseq(1000,12500)
freqseq(bands = "octave")

}
