% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoyance.R
\name{lconvert}
\alias{lconvert}
\title{Conversion curves for equally annoyed people.}
\usage{
lconvert(Lref, ref = "road", type, annoyance = "HA", C = 72,
  coeff = FALSE)
}
\arguments{
\item{Lref}{noise level of the reference source to be converted.}

\item{ref}{type of reference source, 'road' is the default value.}

\item{type}{type of the converted noise level, available source type:
'road', 'aircraft', 'railway', 'wtn'.}

\item{annoyance}{annoyance category. Use the default 'HA' for the percentage of highly annoyed people, 'A' for annoyed and 'LA' for little annoyed.}

\item{C}{cut off value for HA, A, LA definitions. Modify only to use different thresholds in the statistical model. If modified the '\code{annoyance}' parameter is ignored.}
}
\description{
The conversion curves '\code{lconvert}' are linear curves
which convert a noise level of a reference source type
to the noise level of a different source, having the same percentage
of annoyed, highly annoyed or little annoyed people.
}
\details{
The function \code{lconvert} directly convert a noise level
from a reference source to the noise level of a different source.
The default reference source is 'road' traffic noise, but the
function can be used to convert levels between any couple of the
available sources. The conversion is possible for each annoyance
category, 'HA', 'A' or 'LA'.

The conversion curves are presented in the paper of
Fredianelli, Carpita, Licitra in references,
and are based on the studies of Miedema-Oudshoorn and
Janssen et al.
}
\examples{

## conversion at equal \%HA from road noise level to wtn
lconvert(Lref = 70, ref ="road", type = "wtn",
                annoyance = "HA")

## conversion back from wtn to road, in output also the coefficients 
lconvert(Lref = 48.91, ref ="wtn", type = "road",
                 annoyance = "HA", coeff = TRUE)

# plotting conversion curve road-wtn
Lroad <- seq(40,80,1)
Lwtn <- lconvert(Lroad, type = "wtn")

plot(Lwtn~Lroad)

## conversion curves for different sources, with road as reference
Lroad <- seq(40,80,1)
Lwtn <- lconvert(Lroad, type = "wtn")
Lrailway <- lconvert(Lroad, type = "railway")
Laircraft <- lconvert(Lroad, type = "aircraft")

plot(Lrailway~Lroad, ylim=c(30,80),
     ylab = "Lconv [dB(A)]", xlab= "Lref [dB(A)]")
lines(Lwtn~Lroad, lty = 2)
lines(Laircraft~Lroad, lty =3)

}
\references{

}
\seealso{
\code{\link{lannoyed}}, \code{\link{pannoyed}}
}
