% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoyance.R
\name{lden}
\alias{lden}
\title{Day-evening-night noise level computation.}
\usage{
lden(Lday, Levening, Lnight, h_day = 14, h_evening = 2, h_night = 8)
}
\arguments{
\item{Lday}{Lday is the A-weighted long-term average sound level
as defined in ISO 1996-2: 1987,
determined over all the day periods of a year}

\item{Levening}{Levening}

\item{Lnight}{Lnight}

\item{h_day}{Day time period hours}

\item{h_evening}{Evening hours}

\item{h_night}{Night hours}
}
\description{
The function '\code{lden}' computes the day-evening-night noise level
defined by the European Directive END 2002/49/EC.
The function takes as input the Lday, Levening and Lnight noise
levels and the number of hours of each time period.
}
\details{
The default values for the time periods duration consider the
Italian national legislations, with  h_day = 12 hours,
h_evening = 2, h_night = 10,
corrisponding to daytime 06-20, evening 20-22 and nighttime 22-06.
The values can be changed according to each specific national legislation.
}
\examples{

## Lden computation using the default Italian time periods
lden(63,60,60)

## Lden computation using values for a different legislation
lden(63,60,60, h_day=12, h_evening=4, h_night=8)

}
\references{

}
