% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ldiff}
\alias{ldiff}
\title{Difference between two sound pressure levels}
\usage{
ldiff(level1, level2)
}
\arguments{
\item{level1}{sound pressure level}

\item{level2}{sound pressure level to subtract from level1}
}
\description{
\code{ldiff} return the difference of two sound pressure levels
}
\examples{

ldiff(63,60)
}
