% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{leq}
\alias{leq}
\title{Equivalent Continuous Sound Pressure Level}
\usage{
leq(levels, na.rm = FALSE)
}
\arguments{
\item{levels}{a vector of noise levels}

\item{na.rm}{a logical value indicating whether ‘NA’ values should be
ignored before the computation proceeds.}
}
\description{
The function '\code{leq}' return the Equivalent Sound Level of a vector of noise levels measured in the same base time period.
}
\examples{
levels <- c(60, 60, 63, 66)
leq(levels)

levels <- c(60, 60, 63, NA, NA, 66)
leq(levels, na.rm=TRUE)


}
