% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{lsum}
\alias{lsum}
\title{Sum of sound pressure levels}
\usage{
lsum(levels, na.rm = FALSE)
}
\arguments{
\item{levels}{a vector of noise levels}

\item{na.rm}{a logical value indicating whether ‘NA’ values should be
ignored before the computation proceeds.}
}
\description{
\code{lsum} return the energetic sum of a vector of sound pressure levels
}
\examples{

lsum(c(40,40))

}
