% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoyance.R
\name{pannoyed}
\alias{pannoyed}
\title{Miedema-Oudshoorn exposure-response curves.}
\usage{
pannoyed(Lden, type = "road", annoyance = "HA", Lden_max = 80, C = 72)
}
\arguments{
\item{Lden}{day-evening-night noise level.
Can be both a single value or a vector of values.}

\item{type}{available source type:
'road', 'aircraft', 'railway', 'wtn'.}

\item{annoyance}{annoyance category. Use the default 'HA' for the percentage of highly annoyed people, 'A' for annoyed and 'LA' for little annoyed.}

\item{Lden_max}{maximum admitted Lden value for the specific source.
If exceeded the function returns \code{NA}.}

\item{C}{cut off value for HA, A, LA definitions. Modify only to use different thresholds in the statistical model. If modified the '\code{annoyance}' parameter is ignored.}
}
\description{
The function '\code{pannoyed}' computes the percentage of people
highly annoyed, annoyed or little annoyed for different sources,
using the Miedema-Oudshoorn exposure-response curves.
For transportation noise are used the data from
the cumulative analysis by Miedema-Oudshoorn 2001,
reported in references.
For Wind Turbine Noise the data from Janssen 2011 are used.
}
\details{
'\code{pannoyed}' takes as input both a single Lden value or a
vector of Lden values, besides the parameters to choose the
source type and annoyance category.
The function '\code{\link{lannoyed}}' computes
the inverse function of 'pannoyed',
obtaining the Lden value corresponding to a given percentage
of annoyed people '\code{p}'.
To compute the confidence and tolerance interval see the function
'pannoyed_conf'.
}
\examples{

## percentage of highly annoyed HA at Lden = 50dB(A)
pannoyed(50, type = "road", annoyance = "HA")
pannoyed(50, type = "aircraft", annoyance = "HA")
pannoyed(50, type = "railway", annoyance = "HA")
pannoyed(50, type = "wtn", annoyance = "HA")

## road exposure-response curves for HA, A, LA
Lden <- seq(30,70)
pannoyed(Lden, type = "road", annoyance = "HA")
pHA_road <- pannoyed(Lden, type = "road", annoyance = "HA")
pA_road <- pannoyed(Lden, type = "road", annoyance = "A")
pLA_road <- pannoyed(Lden, type = "road", annoyance = "LA")

plot(pLA_road~Lden, type = "l", col="red",
xlab="Lden [dB(A)]", ylab="\%p annoyed")
lines(pA_road~Lden, col="blue")
lines(pHA_road~Lden, col="green")

## comparison of HA exposure-response for different sources
Lden <- seq(30,70)
p_road <- pannoyed(Lden, type = "road", annoyance = "HA")
p_aircraft <- pannoyed(Lden, type = "aircraft", annoyance = "HA")
p_railway <- pannoyed(Lden, type = "railway", annoyance = "HA")
p_wtn <- pannoyed(Lden, type = "wtn", annoyance = "HA")

plot(p_road~Lden, type ="l", col="red",
xlab="Lden [dB(A)]", ylab="\%HA")                  
lines(p_aircraft~Lden, col = "blue")
lines(p_railway~Lden, col = "yellow")
lines(p_wtn~Lden, col= "green")

}
\references{

}
\seealso{
\code{\link{lannoyed}}, \code{\link{pannoyed_conf}}
}
