% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoyance.R
\name{pannoyed_conf}
\alias{pannoyed_conf}
\title{Confidence and tolerance intervals for Miedema-Oudshoorn curves.}
\usage{
pannoyed_conf(Lden, level = 0.95, type = "road", annoyance = "HA",
  tol = FALSE, Lden_max = 80, C = 72)
}
\arguments{
\item{Lden}{day-evening-night noise level.
Can be both a single value or a vector of values.}

\item{level}{confidence level, default: 0.95}

\item{type}{available source type:
'road', 'aircraft', 'railway', 'wtn'.}

\item{annoyance}{annoyance category. Use the default 'HA' for the percentage of highly annoyed people, 'A' for annoyed and 'LA' for little annoyed.}

\item{tol}{if TRUE returns both the confidence and the tolerance
intervals. For wind turbine noise tolerance intervals are not
available at the moment.}

\item{Lden_max}{maximum admitted Lden value for the specific source.
If exceeded the function returns \code{NA}.}

\item{C}{cut off value for HA, A, LA definitions. Modify only to use different thresholds in the statistical model. If modified the '\code{annoyance}' parameter is ignored.}
}
\description{
'\code{pannoyed_conf}' computes the confidence and tolerance
intervals for the Miedema-Oudshoorn exposure-response curve
calculated with '\code{\link{pannoyed}}'.
It returns a dataframe containing the confidence/tolerance intervals
and the exposure curve.
}
